requirements.txt
flask
torch
transformers
pillow
git